Update developer.md
clarifying the following points in https://bestpractices.coreinfrastructure.org/en/criteria/0 so that we can get best practice badge. the docs already exist but this makes them easy to find on the website. 

 `The project website MUST provide information on how to: obtain, provide feedback (as bug reports or enhancements), and contribute to the software. [interact]` 

`The information on how to contribute MUST explain the contribution process (e.g., are pull requests used?) {Met URL} [contribution]`


Signed-off-by: Alexa Griffith  <agriffith50@bloomberg.net>